So there are a bunch of changes in here...
RestMail has replaced Gmail as the account confirmation messager.
Facebook and Google social accounts have been split between those already associated with a provider (facebook and google) and those available for registration (facebook_signup and google_signup).
Repaired some of the helper functions for social sign up and log in.
Changed the behavior of some utility functions to be less dependent on the user.